shared library ============== - Change Energy historizer from double to unsigned long.
Plugin RFXCOM 
=============
- Change some conversion due to the Energy type change.

Plugin RFXLANXPL
================
- Change some conversion due to the Energy type change.

Plugin TeleInfoUSB (In Progress)
==================
- Add the first release of the teleinfo plugin.
- Work with BASE contract.
- Should work for HEURE CREUSE / HEURE PLEINE Contract.
- Tempo contract and some messages : Work in progress.

+++ sources/plugins/RfxLanXpl/xplrules/rfxLanXpl/SensorBasic.cpp
@@ -367,7 +367,7 @@ namespace xplrules { namespace rfxLanXpl {
          if (boost::iequals(msg.getBodyValue(m_keywordType), m_keywordTypeEnergy))
          {
             boost::shared_ptr< shared::plugin::yPluginApi::historization::CEnergy > energy(new shared::plugin::yPluginApi::historization::CEnergy(m_keywordTypeEnergy));
-            energy->set(boost::lexical_cast<double>(msg.getBodyValue(m_keywordCurrent)) * 1000.0); //in xpl messgae data is in kW, convert to W
+            energy->set(boost::lexical_cast<unsigned long>(msg.getBodyValue(m_keywordCurrent)) * 1000); //in xpl messgae data is in kW, convert to W
             data.push_back(energy);
          }
